refactor: Update MeadeResponse class with additional methods for payload handling
